Add multi users mode

+++ app/Http/Controllers/Admin/WorkController.php
@@ -2,14 +2,14 @@
 
 namespace App\Http\Controllers\Admin;
 
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\WorkRequest;
+use Illuminate\Support\Facades\Storage;
 use App\Models\Work;
 use App\Models\Type;
 use App\Models\Technology;
-
+use Illuminate\Support\Facades\Auth;
 
 use function PHPSTORM_META\type;
 
@@ -22,7 +22,14 @@ class WorkController extends Controller
      */
     public function index()
     {
-        $works = Work::paginate(10);
+        if(isset($_GET['search'])){
+            $tosearch = $_GET['search'];
+            $works    = Work::where('user_id', Auth::id())
+                            ->where('title', 'like', "%$tosearch%")
+                            ->paginate(10);
+        }else{
+            $works = Work::where('user_id', Auth::id())->paginate(10);
+        }
 
         return view('admin.works.index', compact('works'));
     }
@@ -48,7 +55,7 @@ class WorkController extends Controller
      */
     public function create(Work $work)
     {
-        $types = Type::all();
+        $types        = Type::all();
         $technologies = Technology::all();
 
         return view('admin.works.create', compact('work', 'types', 'technologies'));
@@ -63,18 +70,18 @@ class WorkController extends Controller
     public function store(WorkRequest $request)
     {
         $form_data = $request->all();
-        $form_data['slug'] = Work::generateSlug($form_data['title']);
+        $form_data['slug']          = Work::generateSlug($form_data['title']);
         $form_data['creation_date'] = date('Y-m-d');
+        $form_data['user_id']       = Auth::id();
 
-        if(array_key_exists('image', $form_data)){
-            $form_data['image'] = Storage::put('uploads/', $form_data['image']);
+        if(array_key_exists('image',$form_data)){
+
             $form_data['image_original_name'] = $request->file('image')->getClientOriginalName();
-            //$form_data['image'] = Storage::put('uploads/', $form_data['image']);
+            $form_data['image'] = Storage::put('uploads/', $form_data['image']);
+
         }
 
-        $new_work = new Work;
-        $new_work->fill($form_data);
-        $new_work->save();
+        $new_work = Work::create($form_data);
 
         if(array_key_exists('technologies', $form_data)){
             $new_work->technologies()->attach($form_data['technologies']);
@@ -89,8 +96,15 @@ class WorkController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Work $work)
+    public function show($id)
     {
+        $work = Work::  where('id', $id)
+                        ->where('user_id', Auth::id())
+                        ->first();
+        if(!$work){
+            abort('404');
+        }
+
         $date = date_create($work->creation_date);
         $date_formatted = date_format($date, 'd-m-Y');
         return view('admin.works.show', compact('work', 'date_formatted'));
@@ -119,6 +133,7 @@ class WorkController extends Controller
     public function update(WorkRequest $request, Work $work)
     {
         $form_data = $request->all();
+
         $date = date_create($work->creation_date);
         $date_formatted = date_format($date, 'd-m-Y');
 
@@ -129,14 +144,20 @@ class WorkController extends Controller
         }
         $form_data['creation_date'] = date('Y-m-d');
 
-        if($request->hasFile('image')) {
+        if(array_key_exists('image',$form_data)){
 
             if($work->image){
-                Storage::disk('public')->delete($work->image);
+                    Storage::disk('public')->delete($work->image);
             }
 
-            $form_data['image_original_name'] = $request->file('image')->getClientOriginalName();
-            $form_data['image'] = Storage::put('uploads/', $form_data['image']);
+                $form_data['image_original_name'] = $request->file('image')->getClientOriginalName();
+                $form_data['image_path'] = Storage::put('uploads/',$form_data['image']);
+        }
+
+        if(array_key_exists('noimage', $form_data) && $work->image) {
+            Storage::disk('public')->delete($work->image);
+            $form_data['image_original_name'] = '';
+            $form_data['image'] = '';
         }
 
         $work->update($form_data);
@@ -164,6 +185,6 @@ class WorkController extends Controller
 
         $work->delete();
 
-        return redirect()->route('admin.works.index')->with('deleted', "Il lavoro: \" $work->title \" è stato eliminato correttamente!");
+        return redirect()->route('admin.works.index')->with('deleted', "Il lavoro: \" $work->title \" è stato eliminato con successi!");
     }
 }
